chore: move assertEventFiredShallow to test helpers
