+++ tests/mocha/test_helpers/events.js
@@ -131,6 +131,49 @@ export function assertEventFired(spy, instanceType, expectedProperties,
 }
 
 /**
+ * Returns a matcher that asserts that the actual object has the same properties
+ * and values (shallowly equated) as the expected object.
+ * @param {!Object} expected The expected set of properties we expect the
+ *    actual object to have.
+ * @return {function(*): boolean} A matcher that returns true if the `actual`
+ *     object has all of the properties of the `expected` param, with the same
+ *     values.
+ */
+function shallowMatch(expected) {
+  return (actual) => {
+    for (const key in expected) {
+      if (actual[key] !== expected[key]) {
+        return false;
+      }
+    }
+    return true;
+  };
+}
+
+/**
+ * Asserts that an event with the given values (shallowly evaluated) was fired.
+ * @param {!SinonSpy|!SinonSpyCall} spy The spy or spy call to use.
+ * @param {function(new:Blockly.Events.Abstract)} instanceType Expected instance
+ *    type of event fired.
+ * @param {!Object<string, *>} expectedProperties Map of of expected properties
+ *    to check on fired event.
+ * @param {string} expectedWorkspaceId Expected workspace id of event fired.
+ * @param {?string=} expectedBlockId Expected block id of event fired.
+ */
+export function assertEventFiredShallow(
+    spy, instanceType, expectedProperties, expectedWorkspaceId, expectedBlockId) {
+  const properties = {
+    ...expectedProperties,
+    workspaceId: expectedWorkspaceId,
+    blockId: expectedBlockId,
+  };
+  sinon.assert.calledWith(
+      spy,
+      sinon.match.instanceOf(instanceType)
+          .and(shallowMatch(properties)));
+}
+
+/**
  * Asserts that an event with the given values was not fired.
  * @param {!SpyCall} spy The spy to use.
  * @param {function(new:Blockly.Events.Abstract)} instanceType Expected instance
